fix issues, refactoring
